# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Cahoots API
  description: API for Cahoots AI Development Team Platform
  version: 0.1.0
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.cahoots.ai
    description: Production server
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    MemberInvite:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
        role:
          type: string
          default: member
          enum: [admin, member]
    MemberUpdate:
      type: object
      required:
        - role
      properties:
        role:
          type: string
          enum: [admin, member]
    MemberResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        full_name:
          type: string
        role:
          type: string
        is_active:
          type: boolean
...
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
...
paths:
...
  /organizations/{organization_id}/members:
    parameters:
      - name: organization_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags:
        - Organizations
      summary: List organization members
      operationId: listMembers
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: List of organization members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MemberResponse'
        '403':
          description: Invalid or missing API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /organizations/{organization_id}/members/invite:
    parameters:
      - name: organization_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags:
        - Organizations
      summary: Invite member to organization
      operationId: inviteMember
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemberInvite'
      responses:
        '201':
          description: Invitation sent successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: User is already a member
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Invalid or missing API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /organizations/{organization_id}/members/{user_id}:
    parameters:
      - name: organization_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    patch:
      tags:
        - Organizations
      summary: Update member role
      operationId: updateMember
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemberUpdate'
      responses:
        '200':
          description: Member role updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberResponse'
        '404':
          description: User not found or not a member
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Organizations
      summary: Remove member from organization
      operationId: removeMember
      security:
        - ApiKeyAuth: []
      responses:
        '204':
          description: Member removed successfully
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...